requirements.txt
amqp==1.4.9
ansible==2.0.2.0
anyjson==0.3.3
Babel==0.9.6
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.4.1
billiard==3.3.0.23
bs4==0.0.1
celery==3.1.23
certifi==2016.2.28
chardet==2.0.1
click==6.6
configobj==4.7.2
cx-Oracle==5.2.1
decorator==3.4.0
ecdsa==0.13
et-xmlfile==1.0.1
fcon==1.1.0
figleaf==0.6.1
Flask==0.11.1
Flask-MySQLdb==0.2.0
happybase==0.9
httplib2==0.7.7
ibm-db==2.0.7
iniparse==0.4
ipython==4.2.1
ipython-genutils==0.1.0
itsdangerous==0.24
javapackages==1.0.0
jdcal==1.3
Jinja2==2.7.2
kitchen==1.1.1
kombu==3.0.35
langtable==0.0.13
lockfile==0.12.2
logging==0.4.9.6
lxml==3.2.1
MarkupSafe==0.11
mysqlclient==1.3.9
numpy==1.12.0
openpyxl==2.4.0
pandas==0.19.2
paramiko==1.15.1
pathlib2==2.1.0
pexpect==4.1.0
pickleshare==0.7.2
PSI==0.3b2
psycopg2==2.6.2
ptyprocess==0.5.1
pyasn1==0.1.6
pycassa==1.11.1
pycrypto==2.6.1
pycurl==7.19.0
pygobject==3.8.2
pygpgme==0.3
PyGreSQL==4.2
PyHDFS==0.1.1
pyhs2==0.6.0
pyliblzma==0.5.3
pymongo==3.4.0
pymssql==2.1.3
PyMySQL==0.7.5
pyparsing==1.5.6
python-augeas==0.4.1
python-dateutil==2.6.0
python-keyczar==0.71rc0
pytz==2016.4
pyudev==0.15
pyxattr==0.5.1
PyYAML==3.10
requests==2.10.0
sasl==0.2.1
simplegeneric==0.8.1
simplejson==3.9.0
singledispatch==3.4.0.3
six==1.9.0
slip==0.4.0
slip.dbus==0.4.0
spark-submit-app-id-wrapper==0.1.0
SQLAlchemy==1.0.13
thrift==0.9.3
tornado==4.3
traitlets==4.2.2
urlgrabber==3.10
vboxapi==1.0
Werkzeug==0.11.11
xlrd==1.0.0
xlutils==2.0.0
xlwt==1.1.2
yum-langpacks==0.4.2
yum-metadata-parser==1.1.4
